準備覆蓋用

+++ CoreMVC5_UsedBookProject/Controllers/AdministratorProfile.cs
@@ -1,46 +1,34 @@
 ﻿using CoreMVC5_UsedBookProject.Data;
-using CoreMVC5_UsedBookProject.Interfaces;
 using CoreMVC5_UsedBookProject.Models;
-using CoreMVC5_UsedBookProject.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Runtime.InteropServices;
+using System.Linq;
 using System.Threading.Tasks;
-using static System.Net.WebRequestMethods;
-
+using CoreMVC5_UsedBookProject.ViewModels;
 namespace CoreMVC5_UsedBookProject.Controllers
 {
-    public class AdministratorHomePageController : Controller
+    [Authorize(Roles = "Administrator")]
+    public class AdministratorProfile : Controller
     {
-        private readonly IWebHostEnvironment _webHostEnvironment;
         private readonly AdminAccountContext _ctx;
 
-        public AdministratorHomePageController(AdminAccountContext ctx)
+        public AdministratorProfile(AdminAccountContext ctx)
         {
             _ctx = ctx;
         }
-        public override void OnActionExecuted(ActionExecutedContext context)
-        {
-            base.OnActionExecuted(context);
-            var NickName = HttpContext.Request.Cookies["NickName"];
-            ViewBag.NickName = NickName;
-            var UserIcon = HttpContext.Request.Cookies["UserIcon"];
-            ViewBag.UserIcon = UserIcon;
-        }
+
         [Authorize]
-        public async Task<IActionResult> AdministratorHomePage()
+        public async Task<IActionResult> AdministratorData()
         {
             var model = await _ctx.Users.ToListAsync();
-
             return View(model);
         }
-        [Authorize(Roles = "Administrator")]
+
+        
         public async Task<IActionResult> AdministratorDetail(string id)
         {
             if (string.IsNullOrEmpty(id))
@@ -53,7 +41,10 @@ namespace CoreMVC5_UsedBookProject.Controllers
                 return new BadRequestObjectResult(msgObject);
             }
 
-            var administratorUser = await _ctx.Users.FirstOrDefaultAsync(m => m.Id == id);
+            //var administratorUser = await _ctx.Users.FirstOrDefaultAsync(m => m.Id == id);
+            var administratorUser = await _ctx.Users.Where(w=>w.Id == id)
+                .Select(g => new AdministratorUserViewModel { Id = id, Name = g.Name, Nickname = g.Nickname,Email = g.Email, PhoneNo = g.PhoneNo })
+                .FirstOrDefaultAsync();
 
             if (administratorUser == null)
             {
@@ -63,96 +54,92 @@ namespace CoreMVC5_UsedBookProject.Controllers
             return View(administratorUser);
         }
 
-        [HttpGet]
-        [Authorize(Roles = "Administrator")]
         public IActionResult AdministratorCreate()
         {
             return View();
         }
+
         [HttpPost]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles = "Administrator")]
-        public async Task<IActionResult>
-            Create([Bind("Name,Nickname,Email,PhoneNo")] AdministratorUser administratorUser)
+        public async Task<IActionResult> AdministratorCreate([Bind("Name,Nickname,Email,PhoneNo")] AdministratorUser administratorUser)
         {
             if (ModelState.IsValid)
             {
                 _ctx.Users.Add(administratorUser);
                 await _ctx.SaveChangesAsync();
-                return RedirectToAction("Index");
+                return RedirectToAction(nameof(AdministratorHomePage));
             }
 
             return View(administratorUser);
         }
 
-        [HttpGet]
-        [Authorize(Roles = "Administrator")]
         public async Task<IActionResult> AdministratorDelete(string id)
         {
             if (id == null)
             {
                 return NotFound();
             }
+
             var administratorUser = await _ctx.Users.FirstOrDefaultAsync(m => m.Id == id);
+
             if (administratorUser == null)
             {
                 return NotFound();
             }
+
             return View(administratorUser);
         }
+
         [HttpPost, ActionName("AdministratorDelete")]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles = "Administrator")]
-        public async Task<IActionResult> DeleteConfirmed(string id)
+        public async Task<IActionResult> AdministratorDeleteConfirmed(string id)
         {
             var administratorUser = await _ctx.Users.FindAsync(id);
+
+            if (administratorUser == null)
+            {
+                return NotFound();
+            }
+
             _ctx.Users.Remove(administratorUser);
             await _ctx.SaveChangesAsync();
-            return RedirectToAction("Index");
+            return RedirectToAction(nameof(AdministratorHomePage));
         }
 
-
-
-
-
-
-
-
-
-
-        [HttpGet]
-        [Authorize(Roles = "Administrator")]
         public async Task<IActionResult> AdministratorEdit(string id)
         {
             if (id == null)
             {
                 return NotFound();
             }
-            var administratorUser = await _ctx.Users.FirstAsync(m => m.Id == id);
+
+            var administratorUser = await _ctx.Users.FirstOrDefaultAsync(m => m.Id == id);
+
             if (administratorUser == null)
             {
                 return NotFound();
             }
+
             return View(administratorUser);
         }
+
         [HttpPost]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles = "Administrator")]
-        public async Task<IActionResult> Edit(string id,
-            [Bind("Name,Nickname,Email,PhoneNo")] AdministratorUser administratorUser)
+        public async Task<IActionResult> AdministratorEdit(string id, [Bind("Id,Name,Nickname,Email,PhoneNo")] AdministratorUser administratorUser)
         {
             if (id != administratorUser.Id)
             {
                 return NotFound();
             }
-            if(ModelState.IsValid)
+
+            if (ModelState.IsValid)
             {
                 try
                 {
                     _ctx.Users.Update(administratorUser);
                     await _ctx.SaveChangesAsync();
                 }
-                catch(DbUpdateConcurrencyException)
+                catch (DbUpdateConcurrencyException)
                 {
                     if (!AdministratorUserExists(administratorUser.Id))
                     {
@@ -163,14 +150,16 @@ namespace CoreMVC5_UsedBookProject.Controllers
                         throw;
                     }
                 }
-                return RedirectToAction("Index");
+
+                return RedirectToAction(nameof(AdministratorHomePage));
             }
+
             return View(administratorUser);
         }
 
         private bool AdministratorUserExists(string id)
         {
-            throw new NotImplementedException();
+            return _ctx.Users.Any(e => e.Id == id);
         }
     }
 }
